requirements.txt
alembic==0.8.5
astral==3.2
attrs==23.1.0
beautifulsoup4==4.4.1
blinker==1.6.2
cachetools==5.5.0
certifi==2023.11.17
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.3
Flask==2.3.2
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gunicorn==19.4.5
h11==0.14.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
Mako==1.0.4
markovify==0.9.4
MarkupSafe==2.1.3
nba_api==1.4.1
nltk==3.4.5
numpy==1.26.3
oauthlib==3.2.2
outcome==1.2.0
pandas==2.0.3
proto-plus==1.24.0
protobuf==4.21.12
psycopg2-binary==2.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==0.5
pytz==2024.1
redis==2.10.5
requests==2.31.0
requests-oauthlib==2.0.0
rq==0.5.6
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.0.12
threadpoolctl==3.2.0
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.4
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==2.0.3
Werkzeug==2.3.6
wsproto==1.2.0
